LayoutTests/fast/dom/select-options-collection-idempotence.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>

description("This test verifies that HTMLSelectElement.options returns the same collection when called repeatedly.");

var selectElement = document.createElement("select");

shouldBe("selectElement.options", "selectElement.options");
shouldBeTrue("selectElement.options === selectElement.options");

</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
