LayoutTests/fast/dom/select-options-collection-idempotence-expected.txt

This test verifies that HTMLSelectElement.options returns the same collection when called repeatedly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS selectElement.options is selectElement.options
PASS selectElement.options === selectElement.options is true
PASS successfullyParsed is true

TEST COMPLETE
